+++ src/blocks.js
@@ -7610,6 +7610,35 @@ RingMorph.prototype.userMenu = function () {
     return RingMorph.uber.userMenu.call(this);
 };
 
+// RingMorph op-sequence analysis
+
+RingMorph.prototype.unwind = function () {
+    // start with the formal parameters and then go backwards
+    return this.inputs()[1].unwind();
+};
+
+RingMorph.prototype.unwindAfter = function (element) {
+    // start with the formal parameters and then go backwards
+    var idx = this.inputs().indexOf(element),
+        current,
+        nxt;
+    if (idx < 1) { // begin of ring
+        if (this.parent instanceof MultiArgMorph ||
+                this.parent instanceof BlockMorph) {
+            nxt = this.parent;
+            current = this;
+        } else if (this.parent instanceof ArgMorph) {
+            nxt = this.parent.parentThatIsA(BlockMorph);
+            current = this.parent;
+        }
+        if (nxt) {
+            return [this].concat(nxt.unwindAfter(current));
+        }
+        return [this];
+    }
+    return this.inputs()[idx - 1].unwind();
+};
+
 // ScriptsMorph ////////////////////////////////////////////////////////
 
 /*
@@ -13882,7 +13911,7 @@ ReporterSlotMorph.prototype.fixLayout = function () {
 
 // ReporterSlotMorph op-sequence analysis
 
-ReporterSlotMorph.prototype.unwind = CommandSlotMorph.prototype.unwind
+ReporterSlotMorph.prototype.unwind = CommandSlotMorph.prototype.unwind;
 
 // RingReporterSlotMorph ///////////////////////////////////////////////////